[cleanup] Move non generated code to seperate file (#9121)
* [codegen] Move non-generated code out of template

* Fix darwin

* Run Codegen
