+++ src/controller/data_model/zap-generated/CHIPClientCallbacks.h
@@ -24,6 +24,7 @@
 #include <app/common/gen/af-structs.h>
 #include <app/util/af-enums.h>
 #include <app/util/attribute-filter.h>
+#include <app/util/im-client-callbacks.h>
 #include <inttypes.h>
 #include <lib/support/FunctionTraits.h>
 #include <lib/support/Span.h>
@@ -31,26 +32,6 @@
 // Note: The IMDefaultResponseCallback is a bridge to the old CallbackMgr before IM is landed, so it still accepts EmberAfStatus
 // instead of IM status code.
 // #6308 should handle IM error code on the application side, either modify this function or remove this.
-bool IMDefaultResponseCallback(const chip::app::Command * commandObj, EmberAfStatus status);
-bool IMReadReportAttributesResponseCallback(const chip::app::ReadClient * apReadClient, const chip::app::ClusterInfo & aPath,
-                                            chip::TLV::TLVReader * apData, chip::Protocols::InteractionModel::ProtocolCode status);
-bool IMWriteResponseCallback(const chip::app::WriteClient * writeClient, EmberAfStatus status);
-
-// Global Response Callbacks
-typedef void (*BooleanAttributeCallback)(void * context, bool value);
-typedef void (*Int8uAttributeCallback)(void * context, uint8_t value);
-typedef void (*Int8sAttributeCallback)(void * context, int8_t value);
-typedef void (*Int16uAttributeCallback)(void * context, uint16_t value);
-typedef void (*Int16sAttributeCallback)(void * context, int16_t value);
-typedef void (*Int32uAttributeCallback)(void * context, uint32_t value);
-typedef void (*Int32sAttributeCallback)(void * context, int32_t value);
-typedef void (*Int64uAttributeCallback)(void * context, uint64_t value);
-typedef void (*Int64sAttributeCallback)(void * context, int64_t value);
-typedef void (*OctetStringAttributeCallback)(void * context, const chip::ByteSpan value);
-typedef void (*CharStringAttributeCallback)(void * context, const chip::ByteSpan value);
-typedef void (*AttributeResponseFilter)(chip::TLV::TLVReader * data, chip::Callback::Cancelable * onSuccess,
-                                        chip::Callback::Cancelable * onFailure);
-
 typedef void (*ReadReportingConfigurationReportedCallback)(void * context, uint16_t minInterval, uint16_t maxInterval);
 typedef void (*ReadReportingConfigurationReceivedCallback)(void * context, uint16_t timeout);
